requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
Faker==38.2.0
inflection==0.5.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
